backend/pyproject.toml
[tool.poetry]
name = "keemu-backend"
version = "0.1.0"
description = "KeeMU Content Intelligence Assistant - Backend API"
authors = ["KeeMU Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"

# Web Framework 
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-multipart = "^0.0.20"

# Database 
sqlalchemy = "^2.0.35"
asyncpg = "^0.30.0"
alembic = "^1.13.3"
psycopg2-binary = "^2.9.10"
pgvector = "^0.3.5"

# Redis & Celery 
redis = "^5.2.0"
celery = {extras = ["redis"], version = "^5.4.0"}
flower = "^2.0.1"

# Authentication 
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
google-auth = "^2.36.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"

# AI/ML 
anthropic = "^0.39.0"
sentence-transformers = "^3.3.0"
torch = "^2.5.0"
openai = "^1.54.0"
tiktoken = "^0.8.0"

# Content Collection 
youtube-transcript-api = "^0.6.2"
google-api-python-client = "^2.150.0"
isodate = "^0.7.2"
praw = "^7.8.1"
feedparser = "^6.0.11"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"
requests = "^2.32.3"
aiohttp = "^3.11.0"

# Blog/RSS Integration - Task 6
fastfeedparser = "^0.3.2"
trafilatura = "^1.12.2"
newspaper4k = "^0.9.3"
readability-lxml = "^0.8.1"

# Email - Latest stable version
sendgrid = "^6.11.0"

# Utilities 
python-decouple = "^3.8"
pydantic = "^2.9.0"
pydantic-settings = "^2.6.0"
python-dateutil = "^2.9.0"
pytz = "^2024.2"

# Logging & Monitoring 
structlog = "^24.4.0"
sentry-sdk = {extras = ["fastapi"], version = "^2.18.0"}

# Data Validation - Latest stable version
email-validator = "^2.2.0"

[tool.poetry.group.dev.dependencies]
# Testing 
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
httpx = "^0.27.0"
faker = "^30.0.0"

# Code Quality 
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.0"
mypy = "^1.13.0"
pylint = "^3.3.0"
pre-commit = "^4.0.0"

# Development 
ipython = "^8.29.0"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.tox
  | dist
  | alembic
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=app --cov-report=html --cov-report=term-missing"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
